test: move rate limiter tests to performance
The tests assume they can do a certain amount of work in a fixed time,
which may not work if we run lots of tests in parallel with xdist.

Move them to performance tests so they run single-threaded.
